Патч #3 Сделано: * убрана заглушка и добавлена фича с RatedSeason

+++ src/main/java/ru/youmiteru/backend/dto/SeasonDto/RatedSeason.java
@@ -4,6 +4,6 @@ public record RatedSeason(
     Long seasonId,
     String seasonName,
     String imageUrl,
-    Double seasonRating
+    int seasonRating
 ) {
 }
